Ensure GPS fixes have monotonically increasing timestamps.
Testing with tethys/missionlogs/2012/20120908_20120920/20120917T025522/201209170255_201209171110.nc4 revealed this error: ValueError: conflicting sizes for dimension 'nudged_time': length 63728 on the data but length 63611 on coordinate 'nudged_time' which I gave to Claude. I told it to keep digging on the problem and it eventually found the initial cause:

There it is! GPS fix 0 has a timestamp after GPS fix 1! The first two GPS fixes are out of order:

GPS fix 0: 2012-09-17T03:05:44
GPS fix 1: 2012-09-17T03:04:52

This is why we now have this commit.